Refactor import statements for consistency: consolidate _Power imports across multiple files
