+++ CityOfBinds/src/content_managers/templates/utils/commands_template.py
@@ -1,5 +1,5 @@
 from .command_factory import _CommandFactory
-from ....game.utils.powers import _Power
+from ....game.utils import _Power
 from ....game.utils.slash_commands import _CommandGroup, CommandGroupConstants
 from .....utils import ListTemplate, Pool
