+++ CityOfBinds/src/game/utils/__init__.py
@@ -1,7 +1,9 @@
 from .comments.comment import _Comment
 from .comments.comment_banner import _CommentBanner
+from .powers.power import _Power
 
 __all__ = [
     "_Comment",
     "_CommentBanner",
+    "_Power",
 ]
