--- CityOfBinds/src/game/utils/powers/__init__.py
@@ -1,3 +1,0 @@
-from .power import _Power
-
-__all__ = ["_Power"]
