+++ tests/test_power.py
@@ -1,5 +1,5 @@
 import pytest
-from CityOfBinds.src.game.utils.powers import _Power
+from CityOfBinds.src.game.utils import _Power
 
 
 class TestInitialization:
